go-controller: fold ovn-k8s-overlay common.go into gateway_init.go
gateway_init is the only thing left so move things back together.
